Update search to reduce rerenders
