+++ src/components/topbar.tsx
@@ -8,7 +8,7 @@ import { TwitchUser } from "./twitch/TwitchUser";
 import { TwitchLoginButton } from "./twitch/TwitchLoginButton";
 import { getColorMode } from "~src/utils/colormode";
 import { Upload } from "react-bootstrap-icons";
-import { GlobalSearch } from "~src/components/search/global-search";
+import { GlobalSearch } from "~src/components/search/global-search.component";
 import { resetSession } from "~src/actions/reset-session.action";
 
 const DarkModeSlider = dynamic(() => import("./dark-mode-slider"), {
